% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfbconfs}
\alias{cfbconfs}
\title{Subdivisions and conferences}
\format{A data frame with 25 rows and 2 variables:
\itemize{
 \item DIV: NCAA Division-I subdivision (FBS, FCS)
 \item CONF: Athletic conference
}}
\usage{
cfbconfs
}
\description{
A dataset containing the conferences of the two NCAA Division-I
subdivisions.
}
\keyword{datasets}

