% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_game.R
\name{drive_score}
\alias{drive_score}
\alias{drive_score_}
\title{Retrieve cumulative scores at end of each drive for a team}
\usage{
drive_score(accordion, side, half)

drive_score_(accordion, side)
}
\arguments{
\item{accordion}{HTML for a team drive "accordion" on
\code{espn.go.com}.}

\item{side}{Team to scrape; either `home` or `away`.}
}
\value{
A vector where elements represent one team's cumulative
  number of points at the end of each drive and all other elements
  are \code{NA}.

A vector the length of the number of plays in a drive, where
  the last element is one team's cumulative number of points at the
  end of the drive and all other elements are \code{NA}.
}
\description{
Retrieve cumulative scores at end of each drive for a team

Retrieve cumulative score at end of drive
}
\keyword{internal}

