% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_game.R
\name{get_game}
\alias{get_game}
\alias{get_game_}
\title{Retrieve the play-by-play for a given game}
\usage{
get_game(ids)

get_game_(id)
}
\arguments{
\item{ids}{Vector of \code{espn.go.com} game identifiers, as either
characters or numerics.}
}
\value{
A list of team information and play-by-play data presented in
  a dataframe. If \code{`teams`} is a vector, a list of lists is
  returned.

A list of team information and play-by-play presented in a
  dataframe.
}
\description{
Scrapes the play-by-play for a given football game.

Non-vectorized play-by-play scraper
}
\examples{
\dontrun{
get_game(c(400852682, 400852737))
}
}
\keyword{internal}

