% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_schedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Retrieve schedule for a given team and season}
\usage{
get_schedule(teams, yrs)
}
\arguments{
\item{teams}{Vector of teams for which to extract schedules. Elements
should match team names available in \code{cfbteams}.}

\item{yrs}{Vector of years for which to extract schedules. Elements
should directly correspond with those passed to \code{`teams`}.}
}
\value{
A list of data frames of the requested schedules.
}
\description{
Scrapes a football schedule and weekly overview for a given team and
season.
}
\examples{
\dontrun{
get_schedule(c('Iowa', 'Stanford'), rep(2015, 2))
get_schedule(rep('Alabama', 3), 2010:2012)
}
}

