% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_schedule.R
\name{get_schedule_}
\alias{get_schedule_}
\title{Non-vectorized schedule scraper}
\usage{
get_schedule_(id, yr)
}
\value{
A data frame of the requested schedule
}
\description{
Non-vectorized schedule scraper
}
\keyword{internal}

