% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fpi.R
\name{getFPI}
\alias{getFPI}
\title{Get the ESPN Football Power Index}
\usage{
getFPI(yr = format(Sys.Date(), "\%Y"), wk = NULL)
}
\arguments{
\item{yr}{Year of the requested rankings}

\item{wk}{Week of season through which the index should be requested}
}
\value{
A data frame of the requested leaderboard
}
\description{
Extracts an NCAAF Football Power Index ranking from ESPN.com and
returns it as a data frame.
}
\examples{
# Capture Week 2 FPI rankings
getFPI(2015, wk = 2)
}

