% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leaders.R
\name{getLeaders}
\alias{getLeaders}
\title{Get NCAAF leaderboards}
\usage{
getLeaders(x = c("total", "downs", "passing", "rushing", "receiving",
  "returning", "kicking", "punting", "defense"), pos = c("offense",
  "defense"), yr = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{x}{Statistics group of the requested rankings}

\item{pos}{Position of the requested rankings}

\item{yr}{Year of the requested rankings}
}
\value{
A data frame of the requested leaderboard
}
\description{
Extracts NCAAF team statistics leaderboard from ESPN.com and returns
it as a data frame.
}
\examples{
# Capture team rushing offense stats for 2013
df <- getLeaders('rushing', yr = 2013)
}

