% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankings.R
\name{getPowerRankings}
\alias{getPowerRankings}
\title{Get NCAAF power rankings}
\usage{
getPowerRankings(wk)
}
\arguments{
\item{wk}{Week of the requested rankings}
}
\value{
A data frame of the requested leaderboard
}
\description{
Extracts a week of NCAAF team power rankings from ESPN.com and
returns it as a data frame.
}
\examples{
# Capture ESPN power rankings for Week 3
getPowerRankings(3)
}

