% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qbr.R
\name{getQBR}
\alias{getQBR}
\title{Get NCAAF quarterback ratings}
\usage{
getQBR(x = c("player-season", "player-week", "player-game", "team-season",
  "alltime-season", "alltime-game"), qlf = TRUE, yr = format(Sys.Date(),
  "\%Y"))
}
\arguments{
\item{x}{Type of the requested rankings}

\item{qlf}{Whether rankings should use qualified ratings}

\item{yr}{Year of the requested rankings}
}
\value{
A data frame of the requested rankings
}
\description{
Extracts a period of NCAAF QBR rankings from ESPN.com and returns it
as a data frame.
}
\examples{
# Capture QBR single game rankings for 2013
getQBR('player-game', yr = 2013)
}

