% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Get season schedule}
\usage{
getSchedule(team, yr = format(Sys.Date(), "\%Y"))
}
\arguments{
\item{team}{Team for which to extract schedule}

\item{yr}{Year of the requested schedule}
}
\value{
A data frame of the requested schedule
}
\description{
Extracts an NCAAF schedule and overview results for a team.
}
\examples{
getSchedule('iowa-hawkeyes', 2013)
}

