% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_game.R
\name{logo_id_}
\alias{logo_id_}
\title{Retrieve the offensive team identifier for a drive}
\usage{
logo_id_(accordion)
}
\arguments{
\item{accordion}{HTML for a team drive "accordion" on
\code{espn.go.com}.}
}
\value{
A team identifier as a character
}
\description{
Retrieve the offensive team identifier for a drive
}
\keyword{internal}

