% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_game.R
\name{parse_pts_}
\alias{parse_pts_}
\title{Adjust a vector of cumulative points to points scored per drive}
\usage{
parse_pts_(v)
}
\arguments{
\item{v}{Vector of cumulative points.}
}
\value{
A vector of points as numerics
}
\description{
Adjust a vector of cumulative points to points scored per drive
}
\keyword{internal}

