% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_game.R
\name{team_info_}
\alias{team_info_}
\title{Retrieve high-level information for a team at game time}
\usage{
team_info_(side)
}
\arguments{
\item{side}{Team to scrape; either `home` or `away`.}
}
\value{
A list of team information
}
\description{
Retrieve high-level information for a team at game time
}
\keyword{internal}

