% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams}
\alias{teams}
\title{Database identifiers for teams}
\format{A data frame with 253 rows and 3 variables:
\itemize{
 \item CONF: Athletic conference
 \item TEAM: NCAA Division-I team
 \item ESPN.ID: espn.go.com team identifier
}}
\usage{
teams
}
\description{
A dataset containing teams and their online database identifiers.
}
\keyword{datasets}

